requirements.txt
asgiref==3.6.0
Django==4.1.5
sqlparse==0.4.3
djangorestframework==3.14.0
